refactor: common module 추출
